+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/smartcontracts/WasmModuleVersion.java
@@ -14,7 +14,6 @@ import java.nio.ByteBuffer;
  * V1 contracts allows for larger state and module size (i.e. the actual code deployed on chain)
  * and they also support _synchronized_ calls between contracts.
  */
-@ToString
 public enum WasmModuleVersion {
 
     /**
@@ -55,5 +54,16 @@ public enum WasmModuleVersion {
 
     byte[] getBytes() {
         return UInt32.from(this.value).getBytes();
+    }
+
+    @Override
+    public String toString() {
+        switch (this) {
+            case V0:
+                return "V0";
+            case V1:
+                return "V1";
+        }
+        throw new IllegalStateException("Unexpected version");
     }
 }